+++ js/modules/form.js
@@ -1,9 +1,9 @@
-import {formOptions} from './form-options.js';
+import {formOptions} from '../options/form-options.js';
 import { previewImage } from './image-preview.js';
 import { initPriceSlider, resetPriceSlider, deactivateSlider, activateSlider } from './price-slider.js';
 import { sendData } from './load.js';
 import { resetFilter } from './filter.js';
-import { onSendSuccess, onSendError } from './state.js';
+import { onSendSuccess, onSendError } from './load-state.js';
 import { setStartPoint } from './map.js';
 
 const form = document.querySelector('.ad-form');
