+++ js/modules/price-slider.js
@@ -1,4 +1,4 @@
-import { sliderOptions } from './slider-options.js';
+import { sliderOptions } from '../options/slider-options.js';
 
 const priceSlider = document.querySelector('.ad-form__slider');
 const priceField = document.querySelector('#price');
