+++ js/options/slider-options.js
